Postgres/requirements.txt
cachetools
gensim
gradio
networkx
nltk
psycopg2-binary
python-dotenv
scikit-learn
spacy
textblob
